🩹 Fix  character -> player 및 prisma model 변경
전체 코드에서 character -> player
prisma model 수정 (rank 테이블, win-draw-lose 위취변경, level 추가, 복합키 설정)

resolves #5
