+++ .gitignore
@@ -41,7 +41,7 @@ build/Release
 # Dependency directories
 node_modules/
 accountClient/
-characterClient/
+playerClient/
 jspm_packages/
 
 # Snowpack dependency directory (https://snowpack.dev/)
