--- resources/character.data.json
@@ -1,244 +1,0 @@
-{
-    "data": [
-        {
-            "name" : "이상헌",
-            "speed": 68,
-            "goal_decision": 32,
-            "shoot_power": 97,
-            "defense": 39,
-            "stamina": 72
-        },
-        {
-            "name" : "무고사",
-            "speed": 34,
-            "goal_decision": 95,
-            "shoot_power": 42,
-            "defense": 40,
-            "stamina": 55
-        },
-        {
-            "name" :"이동경",
-            "speed": 4,
-            "goal_decision": 84,
-            "shoot_power": 11,
-            "defense": 29,
-            "stamina": 27
-        },
-        {
-            "name" :"이승우",
-            "speed": 47,
-            "goal_decision": 14,
-            "shoot_power": 5,
-            "defense": 93,
-            "stamina": 15
-        },
-        {
-            "name" :"정재희",
-            "speed": 20,
-            "goal_decision": 76,
-            "shoot_power": 33,
-            "defense": 24,
-            "stamina": 54
-        },
-        {
-            "name" :"일류첸코",
-            "speed": 87,
-            "goal_decision": 3,
-            "shoot_power": 93,
-            "defense": 21,
-            "stamina": 38
-        },
-        {
-            "name" :"야고",
-            "speed": 34,
-            "goal_decision": 93,
-            "shoot_power": 83,
-            "defense": 82,
-            "stamina": 93
-        },
-        {
-            "name" :"김현욱",
-            "speed": 67,
-            "goal_decision": 11,
-            "shoot_power": 12,
-            "defense": 17,
-            "stamina": 4
-        },
-        {
-            "name" :"이희균",
-            "speed": 61,
-            "goal_decision": 85,
-            "shoot_power": 71,
-            "defense": 35,
-            "stamina": 58
-        },
-        {
-            "name" :"윌리안",
-            "speed": 99,
-            "goal_decision": 3,
-            "shoot_power": 13,
-            "defense": 45,
-            "stamina": 35
-        },
-        {
-            "name" : "루빅손",
-            "speed": 74,
-            "goal_decision": 99,
-            "shoot_power": 53,
-            "defense": 9,
-            "stamina": 33
-        },
-        {
-            "name" :"엄원상",
-            "speed": 80,
-            "goal_decision": 52,
-            "shoot_power": 32,
-            "defense": 41,
-            "stamina": 21
-        },
-        {
-            "name" :"전병관",
-            "speed": 42,
-            "goal_decision": 69,
-            "shoot_power": 97,
-            "defense": 74,
-            "stamina": 53
-        },
-        {
-            "name" : "주민규",
-            "speed": 53,
-            "goal_decision": 87,
-            "shoot_power": 54,
-            "defense": 20,
-            "stamina": 31
-        },
-        {
-            "name" : "양민혁",
-            "speed": 94,
-            "goal_decision": 60,
-            "shoot_power": 54,
-            "defense": 49,
-            "stamina": 59
-        },
-        {
-            "name" :"가브리엘",
-            "speed": 50,
-            "goal_decision": 19,
-            "shoot_power": 52,
-            "defense": 39,
-            "stamina": 29
-        },
-        {
-            "name" :"박용희",
-            "speed": 15,
-            "goal_decision": 22,
-            "shoot_power": 25,
-            "defense": 79,
-            "stamina": 93
-        },
-        {
-            "name" :"마틘 아담",
-            "speed": 76,
-            "goal_decision": 71,
-            "shoot_power": 46,
-            "defense": 36,
-            "stamina": 86
-        },
-        {
-            "name" :"요시노",
-            "speed": 29,
-            "goal_decision": 57,
-            "shoot_power": 39,
-            "defense": 35,
-            "stamina": 31
-        },
-        {
-            "name" :"송민규",
-            "speed": 99,
-            "goal_decision": 22,
-            "shoot_power": 80,
-            "defense": 31,
-            "stamina": 77
-        },
-        {
-            "name" :"제르소",
-            "speed": 99,
-            "goal_decision": 81,
-            "shoot_power": 13,
-            "defense": 64,
-            "stamina": 43
-        },
-        {
-            "name" :"이중민",
-            "speed": 33,
-            "goal_decision": 47,
-            "shoot_power": 66,
-            "defense": 40,
-            "stamina": 3
-        },
-        {
-            "name" :"유리 조나탄",
-            "speed": 72,
-            "goal_decision": 79,
-            "shoot_power": 47,
-            "defense": 72,
-            "stamina": 1
-        },
-        {
-            "name" :"문선민",
-            "speed": 39,
-            "goal_decision": 86,
-            "shoot_power": 27,
-            "defense": 2,
-            "stamina": 27
-        },
-        {
-            "name" :"이건희",
-            "speed": 81,
-            "goal_decision": 91,
-            "shoot_power": 38,
-            "defense": 27,
-            "stamina": 89
-        },
-        {
-            "name" :"이호재",
-            "speed": 38,
-            "goal_decision": 5,
-            "shoot_power": 13,
-            "defense": 84,
-            "stamina": 74
-        },
-        {
-            "name" :"정한민",
-            "speed": 23,
-            "goal_decision": 35,
-            "shoot_power": 30,
-            "defense": 100,
-            "stamina": 15
-        },
-        {
-            "name" :"여홍규",
-            "speed": 90,
-            "goal_decision": 11,
-            "shoot_power": 1,
-            "defense": 5,
-            "stamina": 95
-        },
-        {
-            "name" :"정재상",
-            "speed": 4,
-            "goal_decision": 6,
-            "shoot_power": 41,
-            "defense": 99,
-            "stamina": 80
-        },
-        {
-            "name" :"진성욱",
-            "speed": 16,
-            "goal_decision": 97,
-            "shoot_power": 37,
-            "defense": 40,
-            "stamina": 43
-        }
-    ]
-}
