+++ src/utils/prisma/index.js
@@ -1,5 +1,5 @@
 import { PrismaClient as accountClient } from "../../../prisma/accountClient/default.js";
-import { PrismaClient as characterClient } from "../../../prisma/characterClient/default.js";
+import { PrismaClient as playerClient } from "../../../prisma/playerClient/default.js";
 
 export const accountPrisma = new accountClient({
   log: ["query", "info", "warn", "error"],
@@ -7,7 +7,7 @@ export const accountPrisma = new accountClient({
   errorFormat: "pretty",
 });
 
-export const characterPrisma = new characterClient({
+export const playerPrisma = new playerClient({
   log: ["query", "info", "warn", "error"],
 
   errorFormat: "pretty",
